Consolidate PV UGens
